feat: instantiate connection using a singleton
